feat: Add API docs hook and render list of modules for POC API page
